feat: diary 이미지 업로드 기능 구현중

+++ backend/src/main/java/com/wetness/model/dto/request/GameResultReqDto.java
@@ -9,7 +9,7 @@ public class GameResultReqDto {
     Long userId;
     int[] terminateDate;
     Long gameId;
-    Long workoutId;
+    int workoutId;
     double score;
     int rank;
 }
